Added missing success test assertion Also removed Runscope and used Mirage instead (Runscope wasn't actively being used without the missing assertion), as Runscope no longer has a free plan.
